+++ source/ignore_files_based_on_mediainfo_data/changelog.md
@@ -1,4 +1,7 @@
 
+**<span style="color:#56adda">0.0.4</span>**
+- remove data['add_file_to_pending_tasks'] = True from file test when the test for metadata fails allowing subsequent plugins to test if file should be added to queue
+
 **<span style="color:#56adda">0.0.3</span>**
 - Use a different jsonata library that provides better stability                 
 - improve consistency across all unmanic plugins by using same jsonata library